Update dependencies and fix UI issues
