+++ lib/feature/authentication/login/view/courier_login_page_router_button.dart
@@ -5,18 +5,11 @@ final class _CourierLoginPageRouterButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          flex: 8,
-          child: MainElevatedButton(
-            onPressed: () {
-              context.router.replaceNamed(RoutePath.courierLoginScreen.path);
-            },
-            child: Text(LocaleKeys.mainText_courierLogin.tr()),
-          ),
-        ),
-      ],
+    return MainElevatedButton(
+      onPressed: () {
+        context.router.replaceNamed(RoutePath.courierLoginScreen.path);
+      },
+      child: Text(LocaleKeys.mainText_courierLogin.tr()),
     );
   }
 }
