+++ lib/product/core/base/mixin/headers_mixin.dart
@@ -1,5 +1,23 @@
 mixin HeadersMixin {
-  String? getCookie(Map<String, List<String>>? map) {
+  String? getCookie(Map<String, List<String>>? map, {CookieTypes? type}) {
+    if (type == CookieTypes.setCookie) {
+      return _getSetCookie(map);
+    } else if (type == CookieTypes.authenticationToken) {
+      return _getAuthenticationToken(map);
+    }
+    return null;
+  }
+
+  static String? _getAuthenticationToken(Map<String, List<String>>? map) {
+    if (map == null) return null;
+    final cookie = map['authentication']?.first;
+    if (cookie == null) return null;
+    final cookieSplit = cookie.split(';');
+    final cookieString = cookieSplit[0].split('=');
+    return cookieString[1];
+  }
+
+  static String? _getSetCookie(Map<String, List<String>>? map) {
     if (map == null) return null;
     final cookie = map['set-cookie']?.first;
     if (cookie == null) return null;
@@ -8,3 +26,5 @@ mixin HeadersMixin {
     return cookieString[1];
   }
 }
+
+enum CookieTypes { setCookie, authenticationToken }
