+++ pubspec.yaml
@@ -68,9 +68,11 @@ dependencies:
   get_it: ^7.6.4
   auto_route: ^7.8.4
   intl: ^0.19.0
+  package_info_plus: ^5.0.1
 
 dependency_overrides:
   intl: ^0.19.0
+  package_info_plus: ^5.0.1
   
 dev_dependencies:
   flutter_test:
